[GEOS-9258] core integration
