+++ doc/en/user/source/configuration/index.rst
@@ -4,6 +4,7 @@ Server configuration
 ====================
 
 This page will detail how to set various configuration options in GeoServer as well as test out the services in GeoServer.
+It also will describe how to view the GeoServer instance status.
 
 .. toctree::
    :maxdepth: 1
