+++ src/extension/pom.xml
@@ -362,12 +362,6 @@
         <module>authkey</module>
       </modules>
     </profile>
-    <profile>
-      <id>status-monitoring</id>
-      <modules>
-        <module>status-monitoring</module>
-      </modules>
-    </profile>
 
     <!-- allExtensions profile, includes all modules -->
     <profile>
@@ -413,7 +407,6 @@
         <module>geofence</module>
         <module>geofence-server</module>
         <module>authkey</module>
-        <module>status-monitoring</module>
       </modules>
     </profile>
 
@@ -461,7 +454,6 @@
         <module>geofence</module>
         <module>geofence-server</module>
         <module>authkey</module>
-        <module>status-monitoring</module>
       </modules>
     </profile>
   </profiles>
