--- src/extension/status-monitoring/pom.xml
@@ -1,67 +1,0 @@
-<?xml version="1.0" encoding="ISO-8859-1"?>
-<!-- Copyright (C) 2017 - Open Source Geospatial Foundation. All rights reserved. 
-	This code is licensed under the GPL 2.0 license, available at the root application 
-	directory. -->
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
-	<modelVersion>4.0.0</modelVersion>
-
-  <parent>
-    <groupId>org.geoserver</groupId>
-    <artifactId>extension</artifactId>
-    <version>2.16-SNAPSHOT</version>
-  </parent>
-
-	<groupId>org.geoserver.extension</groupId>
-	<artifactId>gs-status-monitoring</artifactId>
-	<packaging>jar</packaging>
-	<name>Status Monitoring Extension</name>
-
-	<dependencies>
-		<dependency>
-			<groupId>org.geoserver.web</groupId>
-			<artifactId>gs-web-core</artifactId>
-		</dependency>
-		<dependency>
-			<groupId>org.geoserver</groupId>
-			<artifactId>gs-rest</artifactId>
-		</dependency>
-		<dependency>
-			<groupId>com.github.oshi</groupId>
-			<artifactId>oshi-core</artifactId>
-		</dependency>
-		<dependency>
-			<groupId>com.fasterxml.jackson.core</groupId>
-			<artifactId>jackson-databind</artifactId>
-		</dependency>
-		<dependency>
-			<groupId>org.apache.wicket</groupId>
-			<artifactId>wicket-datetime</artifactId>
-			<version>${wicket.version}</version>
-		</dependency>
-		<dependency>
-			<groupId>joda-time</groupId>
-			<artifactId>joda-time</artifactId>
-		</dependency>
-		<!-- Test dependencies -->
-		<dependency>
-			<groupId>org.geoserver</groupId>
-			<artifactId>gs-main</artifactId>
-			<classifier>tests</classifier>
-			<scope>test</scope>
-		</dependency>
-		<dependency>
-			<groupId>org.geoserver.web</groupId>
-			<artifactId>gs-web-core</artifactId>
-			<classifier>tests</classifier>
-			<scope>test</scope>
-			<version>${gs.version}</version>
-		</dependency>
-		<dependency>
-			<groupId>org.springframework</groupId>
-			<artifactId>spring-test</artifactId>
-			<scope>test</scope>
-		</dependency>
-	</dependencies>
-
-</project>
