--- src/extension/status-monitoring/src/main/resources/GeoServerApplication_ko.properties
@@ -1,3 +1,0 @@
-MonitoringPanel.title=\ubaa8\ub2c8\ud130\ub9c1
-SystemStatusMonitorPanel.updated=\uac31\uc2e0: 
-SystemStatusMonitorPanel.info=\uc815\ubcf4
