--- src/extension/status-monitoring/src/main/resources/applicationContext.xml
@@ -1,24 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<beans xmlns="http://www.springframework.org/schema/beans"
-       xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-       xmlns:context="http://www.springframework.org/schema/context" xsi:schemaLocation="
-        http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans.xsd
-        http://www.springframework.org/schema/context http://www.springframework.org/schema/context/spring-context.xsd">
-  <!--
-    Copyright 2017 Open Source Geospatial Foundation. All rights reserved.
-    This code is licensed under the GPL 2.0 license, available at the root application
-    directory.
-  -->
-
-
-  <context:component-scan base-package="org.geoserver.status.monitoring.rest"/>
-  <context:component-scan base-package="org.geoserver.status.monitoring.collector"/>
-
-  <bean id="systemStatusMonitorTab" class="org.geoserver.status.monitoring.web.SystemStatusTab"/>
-
-  <bean class="org.geoserver.platform.ModuleStatusImpl">
-    <constructor-arg index="0" value="gs-status-monitoring"/>
-    <constructor-arg index="1" value="GeoServer System Monitoring"/>
-  </bean>
-
-</beans>
