+++ src/pom.xml
@@ -1571,7 +1571,6 @@
       <descriptor>release/ext-geofence.xml</descriptor>
       <descriptor>release/ext-geofence-server.xml</descriptor>
       <descriptor>release/ext-authkey.xml</descriptor>
-      <descriptor>release/ext-status-monitoring.xml</descriptor>
      </descriptors>
      <finalName>geoserver-2.16-SNAPSHOT</finalName>
      <outputDirectory>${project.build.directory}/release</outputDirectory>
