+++ src/release/pom.xml
@@ -368,11 +368,6 @@
      <artifactId>gs-authkey</artifactId>
      <version>${project.version}</version>
    </dependency>
-   <dependency>
-     <groupId>org.geoserver.extension</groupId>
-     <artifactId>gs-status-monitoring</artifactId>
-     <version>${project.version}</version>
-    </dependency>
  </dependencies>
  <build>
   <plugins>
