+++ src/restconfig/src/main/java/org/geoserver/rest/system/status/ftl-templates/system-status.ftl
