+++ src/restconfig/src/main/resources/applicationContext.xml
@@ -16,6 +16,7 @@
   <context:component-scan base-package="org.geoserver.security.rest"/>
   <context:component-scan base-package="org.geoserver.catalog.rest"/>
   <context:component-scan base-package="org.geoserver.service.rest"/>
+  <context:component-scan base-package="org.geoserver.rest.system.status"/>
   <!-- XStream persister initializer -->
   <bean id="restConfigXStreamPersister" class="org.geoserver.rest.security.RestConfigXStreamPersister" />
