+++ src/web/core/pom.xml
@@ -138,6 +138,15 @@
     <groupId>org.geotools</groupId>
     <artifactId>gt-geopkg</artifactId>
    </dependency>
+	<dependency>
+		<groupId>org.apache.wicket</groupId>
+		<artifactId>wicket-datetime</artifactId>
+		<version>${wicket.version}</version>
+	</dependency>
+	<dependency>
+		<groupId>joda-time</groupId>
+		<artifactId>joda-time</artifactId>
+	</dependency>
    <dependency>
     <groupId>org.geotools.jdbc</groupId>
     <artifactId>gt-jdbc-h2</artifactId>
