+++ src/web/core/src/main/java/org/geoserver/web/system/status/SystemStatusMonitorPanel.html
