+++ src/web/core/src/main/resources/GeoServerApplication.properties
@@ -1071,3 +1071,63 @@ ServiceLayerConfigurationPanel.layerSettings=Layer Settings
 ServiceLayerConfigurationPanel.config=Selectively enable services for layer
 DisabledServicesPalette.availableHeader=Enabled Services
 DisabledServicesPalette.selectedHeader=Disabled Services
+
+# System status names
+MonitoringPanel.title=System Status
+SystemStatusMonitorPanel.updated=Updated at
+SystemStatusMonitorPanel.info=Info
+SystemStatusMonitorPanel.operating_system-operating_system=Operating system
+SystemStatusMonitorPanel.system_uptime-system_uptime=Uptime
+SystemStatusMonitorPanel.system_average_load_1-system_average_load_1=System average load 1 minute
+SystemStatusMonitorPanel.system_average_load_5-system_average_load_5=System average load 5 minutes
+SystemStatusMonitorPanel.system_average_load_15-system_average_load_15=System average load 15 minutes
+SystemStatusMonitorPanel.physical_cpus-physical_cpus=Number of physical CPUs
+SystemStatusMonitorPanel.logical_cpus-logical_cpus=Number of logical CPUs
+SystemStatusMonitorPanel.running_process-running_process=Number of running process
+SystemStatusMonitorPanel.running_threads-running_threads=Number of running threads
+SystemStatusMonitorPanel.cpu_load-cpu_load=CPU load average
+SystemStatusMonitorPanel.per_cpu_load-cpu_1=CPU 1 load
+SystemStatusMonitorPanel.per_cpu_load-cpu_2=CPU 2 load
+SystemStatusMonitorPanel.per_cpu_load-cpu_3=CPU 3 load
+SystemStatusMonitorPanel.per_cpu_load-cpu_4=CPU 4 load
+SystemStatusMonitorPanel.per_cpu_load-cpu_5=CPU 5 load
+SystemStatusMonitorPanel.per_cpu_load-cpu_6=CPU 6 load
+SystemStatusMonitorPanel.per_cpu_load-cpu_7=CPU 7 load
+SystemStatusMonitorPanel.per_cpu_load-cpu_8=CPU 8 load
+SystemStatusMonitorPanel.memory_used-memory_used=Used physical memory 
+SystemStatusMonitorPanel.memory_total-memory_total=Total physical memory 
+SystemStatusMonitorPanel.memory_free-memory_free=Free physical memory
+SystemStatusMonitorPanel.swap_used-swap_used=Used swap memory
+SystemStatusMonitorPanel.swap_total-swap_total=Total swap memory
+SystemStatusMonitorPanel.swap_free-swap_free=Free swap memory
+SystemStatusMonitorPanel.file_system_total_usage-file_system_total_usage=File system usage
+SystemStatusMonitorPanel.partition_used-/=Partition [/] used space
+SystemStatusMonitorPanel.partition_total-/=Partition [/] total space
+SystemStatusMonitorPanel.partition_free-/=Partition [/] free space
+SystemStatusMonitorPanel.partition_used-/dev/loop1=Partition [/dev/loop1] used space
+SystemStatusMonitorPanel.partition_total-/dev/loop1=Partition [/dev/loop1] total space
+SystemStatusMonitorPanel.partition_free-/dev/loop1=Partition [/dev/loop1] free space
+SystemStatusMonitorPanel.partition_used-/dev/loop2=Partition [/dev/loop2] used space
+SystemStatusMonitorPanel.partition_total-/dev/loop2=Partition [/dev/loop2] total space
+SystemStatusMonitorPanel.partition_free-/dev/loop2=Partition [/dev/loop2] free space
+SystemStatusMonitorPanel.partition_used-/dev/loop3=Partition [/dev/loop3] used space
+SystemStatusMonitorPanel.partition_total-/dev/loop3=Partition [/dev/loop3] total space
+SystemStatusMonitorPanel.partition_free-/dev/loop3=Partition [/dev/loop3] free space
+SystemStatusMonitorPanel.partition_used-/dev/sda1=Partition [/dev/sda1] used space
+SystemStatusMonitorPanel.partition_total-/dev/sda1=Partition [/dev/sda1] total space
+SystemStatusMonitorPanel.partition_free-/dev/sda1=Partition [/dev/sda1] free space
+SystemStatusMonitorPanel.partition_used-/dev/loop4=Partition [/dev/loop4] used space
+SystemStatusMonitorPanel.partition_total-/dev/loop4=Partition [/dev/loop4] total space
+SystemStatusMonitorPanel.partition_free-/dev/loop4=Partition [/dev/loop4] free space
+SystemStatusMonitorPanel.network_interfaces_send-network_interfaces_send=Network interfaces send
+SystemStatusMonitorPanel.network_interfaces_received-network_interfaces_received=Network interfaces received
+SystemStatusMonitorPanel.network_interface_send-docker0=Network interface [docker0] send
+SystemStatusMonitorPanel.network_interface_received-docker0=Network interface [docker0] received
+SystemStatusMonitorPanel.network_interface_send-wlp8s0=Network interface [wlp8s0] send
+SystemStatusMonitorPanel.network_interface_received-wlp8s0=Network interface [wlp8s0] received
+SystemStatusMonitorPanel.temperature-temperature=CPU temperature
+SystemStatusMonitorPanel.voltage-voltage=CPU voltage
+SystemStatusMonitorPanel.fan_speed-fan_speed=Fan speed
+SystemStatusMonitorPanel.geoserver_cpu_usage-geoserver_cpu_usage=GeoServer CPU usage
+SystemStatusMonitorPanel.geoserver_threads-geoserver_threads=GeoServer threads
+SystemStatusMonitorPanel.geoserver_jvm_memory_usage-geoserver_jvm_memory_usage=GeoServer JVM memory usage
