+++ src/web/core/src/main/resources/GeoServerApplication_ko.properties
@@ -1025,3 +1025,9 @@ ServiceLayerConfigurationPanel.layerSettings=\ub808\uc774\uc5b4 \uc124\uc815
 ServiceLayerConfigurationPanel.config=\ub808\uc774\uc5b4 \uc11c\ube44\uc2a4\ub97c \uc120\ud0dd\uc801\uc73c\ub85c \ud65c\uc131\ud654
 DisabledServicesPalette.availableHeader=\uc11c\ube44\uc2a4 \ud65c\uc131\ud654
 DisabledServicesPalette.selectedHeader=\uc11c\ube44\uc2a4 \ube44\ud65c\uc131\ud654
+
+# System status names
+MonitoringPanel.title=\ubaa8\ub2c8\ud130\ub9c1
+SystemStatusMonitorPanel.updated=\uac31\uc2e0: 
+SystemStatusMonitorPanel.info=\uc815\ubcf4
+
